test(typeorm-legacy): fix mysql tests
