+++ packages/typeorm-legacy/tests/mysql/schema.sql
